+++ tests/aws/services/sns/test_sns.validation.json
@@ -1,40 +1,4 @@
 {
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_exists_filter_policy": {
-    "last_validated_date": "2023-11-09T20:04:02+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_exists_filter_policy_attributes_array": {
-    "last_validated_date": "2024-04-17T18:10:45+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_filter_policy": {
-    "last_validated_date": "2024-01-25T18:07:57+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_filter_policy_for_batch": {
-    "last_validated_date": "2023-11-09T20:01:32+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_filter_policy_on_message_body[False]": {
-    "last_validated_date": "2023-11-09T19:58:42+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_filter_policy_on_message_body[True]": {
-    "last_validated_date": "2023-11-09T19:58:29+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_filter_policy_on_message_body_array_attributes": {
-    "last_validated_date": "2024-04-17T21:33:09+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_filter_policy_on_message_body_array_of_object_attributes": {
-    "last_validated_date": "2024-04-17T21:32:40+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_filter_policy_on_message_body_dot_attribute": {
-    "last_validated_date": "2023-11-09T17:50:59+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_set_subscription_filter_policy_scope": {
-    "last_validated_date": "2023-08-24T22:15:12+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_sub_filter_policy_nested_property": {
-    "last_validated_date": "2023-08-24T22:15:14+00:00"
-  },
-  "tests/aws/services/sns/test_sns.py::TestSNSFilter::test_sub_filter_policy_nested_property_constraints": {
-    "last_validated_date": "2023-08-24T22:18:18+00:00"
-  },
   "tests/aws/services/sns/test_sns.py::TestSNSPublishCrud::test_empty_sns_message": {
     "last_validated_date": "2023-08-24T20:31:48+00:00"
   },
